♻️ extract ErrorSplash from NotImplemented/NotFoundPage
